Add separate DAO for Coupon Entity #18
- Move coupon by coupon name to CouponDao
