+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/dao/OrderDao.java
@@ -13,24 +13,4 @@ public class OrderDao {
 
     @PersistenceContext
     private EntityManager entityManager;
-
-    /**
-     * Retrieves the Coupon Entity matched with the coupon name
-     *
-     * @param couponName The coupon name to be searched in Database
-     * @return The Coupon Entity matched with the coupon name
-     */
-    public CouponEntity getCouponByCouponName(String couponName) {
-        try {
-            // Reading as List, if there are several records matching with the same coupon, take the first coupon record
-            List<CouponEntity> couponList = entityManager.createNamedQuery("couponByCouponName", CouponEntity.class).setParameter("couponName", couponName).getResultList();
-            if (couponList != null && !couponList.isEmpty()) {
-                return couponList.get(0);
-            } else {
-                return null;
-            }
-        } catch (NoResultException e) {
-            return null;
-        }
-    }
 }
